# target/classes/swagger/adexchange.yaml
...
swagger: '2.0'
info:
  title: TechM Arena Ad Exchange API
  description: API specification for TechM Arena Ad Exchange
  version: "1.0.0"
host: api.test.com
schemes:
  - http
  - https
basePath: /v1
produces:
  - application/json
paths:
...
  /digitalProperty:
    get:
      description: |
        List of DigitalProperty
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - Publisher
      responses:
        200:
          description: List of DigitalProperty
          schema:
            $ref: '#/definitions/DigitalPropertyList'          
    post:
      description: Creates a new DigitalProperty.
      operationId: createDigitalProperty
      parameters:
        - name: digitalProperty
          in: body
          description: DigitalProperty to be created
          required: true
          schema:
            $ref: '#/definitions/DigitalProperty'
      responses:
        '200':
          description: created DigitalProperty
          schema:
            $ref: '#/definitions/DigitalProperty'
    put:
      description: Updates the DigitalProperty.
      operationId: updateDigitalProperty
      parameters:
        - name: digitalProperty
          in: body
          description: DigitalProperty to be updated
          required: true
          schema:
            $ref: '#/definitions/DigitalProperty'
      responses:
        '200':
          description: updated DigitalProperty
          schema:
            $ref: '#/definitions/DigitalProperty'
  /digitalProperty/{id}:
    get:
      description: Returns DigitalProperty for the given id
      operationId: getDigitalProperty
      parameters:
        - name: id
          in: path
          description: ID of DigitalProperty
          required: true
          type: integer
      responses:
        '200':
          description: DigitalProperty response
          schema:
            $ref: '#/definitions/DigitalProperty'
    delete:
      description: deletes a single DigitalProperty based on the ID supplied
      operationId: deleteDigitalProperty
      parameters:
        - name: id
          in: path
          description: ID of DigitalProperty to delete
          required: true
          type: integer
      responses:
        '200':
          description: DigitalProperty deleted
  /digitalProperty/search:
    get:
      description: Search DigitalProperty
      operationId: searchDigitalProperty
      parameters:
        - name: propertyName
          in: query
          description: propertyName
          required: false
          type: string
        - name: description
          in: query
          description: description
          required: false
          type: string
        - name: type
          in: query
          description: type
          required: false
          type: string
        - name: domainURL
          in: query
          description: domainURL
          required: false
          type: string
        - name: publisherId
          in: query
          description: publisherId
          required: false
          type: string
        - name: IABCategory
          in: query
          description: IABCategory
          required: false
          type: string
        - name: IABSubCategory
          in: query
          description: IABSubCategory
          required: false
          type: string
        - name: country
          in: query
          description: country
          required: false
          type: string
        - name: language
          in: query
          description: language
          required: false
          type: string
        - name: status
          in: query
          description: status
          required: false
          type: integer
        - name: createdTimeStart
          in: query
          description: createdTimeStart
          required: false
          type: string
        - name: createdTimeEnd
          in: query
          description: createdTimeEnd
          required: false
          type: string
        - name: updatedTimeStart
          in: query
          description: updatedTimeStart
          required: false
          type: string
        - name: updatedTimeEnd
          in: query
          description: updatedTimeEnd
          required: false
          type: string
        - name: createdByUser
          in: query
          description: createdByUser
          required: false
          type: string
        - name: updatedByUser
          in: query
          description: updatedByUser
          required: false
          type: string
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of DigitalProperty
          schema:
            $ref: '#/definitions/DigitalPropertyList'
  /digitalProperty/search/keyword:
    get:
      description: Search DigitalProperty by keyword
      operationId: searchDigitalPropertyByKeyword
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of DigitalProperty
          schema:
            $ref: '#/definitions/DigitalPropertyList'            
...
definitions:
...
  DigitalProperty:
    type: object
    properties:
      id:
        type: integer
        description: Id
      propertyName:
        type: string
      description:
        type: string
      type:
        type: string
      domainURL:
        type: string
      publisherId:
        type: string
      IABCategory:
        type: string
      IABSubCategory:
        type: string
      country:
        type: string
      language:
        type: string
      status:
        type: integer
      createdTime:
        type: string
        format: date
      updatedTime:
        type: string
        format: date
      createdByUser:
        type: string
      updatedByUser:
        type: string
  DigitalPropertyList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      digitalPropertys:
        type: array
        items:
          $ref: '#/definitions/DigitalProperty'      
...